+++ packages/cli/example/hello.js
@@ -3,16 +3,15 @@ const window = 10
 const settings = {}
 
 function run() {
-  let close5 = this.assets("AAPL", 5, { prop: 'close', daysAgo: 0 })
 
-  this.print(close5)
+  for (const symbol of assets) {
+    this.buy(this.asset(symbol), 1)
+  }
 
-  let sma5 = this.ema(5, "AAPL")
-  this.print(`AAPL sma => ${sma5}`)
-
-  // for (const symbol of assets) {
-  //   this.buy(this.asset(symbol), 1)
-  // }
+  this.print("price AAPL", this.asset("AAPL").close)
+  this.print("cash:", this.getCash())
+  this.print("total capital:", this.getTotalCapital())
+  this.print("positions:", this.getPositions())
 
 }