+++ packages/cli/src/api/code.ts
@@ -11,12 +11,16 @@ export default () => {
    * @param bars 
    * @returns 
    */
-  const runZpCode = (code: string, bars) => {
+  const runZpCode = (code: string, bars, inputs: any = {}) => {
     const start = performance.now()
     
     const zpEnv = new Env({ bars })
+    zpEnv.bind('inputs', inputs)
     zpEnv.bind('barIndex', 1)
 
+    zpEnv.call('setCash', inputs.initialCapital ?? 10_000)
+    zpEnv.call('setPositions', inputs.openPositions ?? [])
+    
     const result = evalCode(zpEnv, code)
     const stop = performance.now()
     const inSeconds = (stop - start) / 1000
@@ -29,9 +33,15 @@ export default () => {
     }
   }
 
-  const runJsCode = (code: string, bars) => {
+  const runJsCode = (code: string, bars, inputs: any = {}) => {
     const start = performance.now()
     const env = createJsEnv(bars)
+
+    env.inputs = inputs
+    env.barIndex = 1
+
+    env.setCash(inputs.initialCapital ?? 10_000)
+    env.setPositions(inputs.openPositions ?? [])
 
     const execFunc = new Function(code)
     const { run } = execFunc()
@@ -51,13 +61,13 @@ export default () => {
     }
   }
 
-  const runCode = (code: string, lang: string, bars: any) => {
+  const runCode = (code: string, lang: string, bars: any, inputs: any = {}) => {
     switch (lang) {
       case 'js':
-        return runJsCode(code, bars)
+        return runJsCode(code, bars, inputs)
 
       case 'zp':
-        return runZpCode(code, bars)
+        return runZpCode(code, bars, inputs)
 
       default:
         throw new Error('Invalid file extension. It should be .js or .zp')
@@ -70,12 +80,14 @@ export default () => {
    * @param openPositions 
    * @returns 
    */
-  const getZpSymbols = (code: string, openPositions: any = []) => {
+  const getZpSymbols = (code: string, openPositions: any = [], inputs: any = {}) => {
     const metaEnv = new Env({ isMeta: true })
     metaEnv.bind('barIndex', 1)
+    metaEnv.bind('inputs', inputs)
+    metaEnv.call('setCash', inputs.initialCapital ?? 10_000)
 
     // Set positions
-    metaEnv.call('setPositions', [openPositions])
+    metaEnv.call('setPositions', [...openPositions])
 
     evalCode(metaEnv, code)
 
@@ -89,20 +101,25 @@ export default () => {
     return { symbols, maxWindow, settings }
   }
 
-  function getJsSymbols (code: string, openPositions: any = []) {
+  function getJsSymbols (code: string, openPositions: any = [], inputs: any = {}) {
     const execFunc = new Function(code)
     const res = execFunc()
+    const symbols = [
+      ...res.assets,
+      ...openPositions.map(p => p.symbol),
+      ...(inputs.openPositions?.map(p => p.symbol) ?? [])
+    ]
 
-    return { symbols: res.assets, maxWindow: res.window ?? 1, settings: res.settings ?? {} }
+    return { symbols, maxWindow: res.window ?? 1, settings: res.settings ?? {} }
   }
 
-  const getSymbols = (code: string, lang: string, openPositions: any) => {
+  const getSymbols = (code: string, lang: string, openPositions: any, inputs: any = {}) => {
     switch (lang) {
       case 'js':
-        return getJsSymbols(code, openPositions)
+        return getJsSymbols(code, openPositions, inputs)
 
       case 'zp':
-        return getZpSymbols(code, openPositions)
+        return getZpSymbols(code, openPositions, inputs)
 
       default:
         throw new Error('Invalid file extension. It should be .js or .zp')