+++ packages/cli/src/commands/execute.ts
@@ -25,10 +25,10 @@ export default () => {
         const extension = path.extname(file)
         const lang = extension === '.js' ? 'js' : 'zp'
         const code = api.code().readCode(file)
-        const requirements = api.code().getSymbols(code, lang, [])
+        const requirements = api.code().getSymbols(code, lang, config.inputs?.openPositions ?? [], config.inputs ?? {})
 
         const bars = await api.data(config).downloadBars(requirements.symbols, requirements.maxWindow, requirements?.settings?.timeframe ?? 1440, opts.date)
-        let result = api.code().runCode(code, lang, bars)
+        let result = api.code().runCode(code, lang, bars, config.inputs ?? {})
 
         console.log(`${clc.green('✔ Success:')} Code was executed successfully`)
         console.log(`${clc.green('✔ Execution time:')} ${clc.bold(result.time.toFixed(2))} seconds\n`)